separando arquivos e criando os tipos restantes
